requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
Authlib==1.6.9
babel==2.18.0
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
boto3==1.42.83
botocore==1.42.83
certifi==2026.2.25
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.6
click==8.2.1
cryptography==46.0.6
distlib==0.4.0
dparse==0.6.4
exceptiongroup==1.3.1
filelock==3.25.2
Flask==3.1.3
Flask-Admin==2.0.2
Flask-BasicAuth==0.2.0
Flask-SQLAlchemy==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.15
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==4.3.0
mdurl==0.1.2
nltk==3.9.4
packaging==26.0
pillow==12.2.0
platformdirs==4.4.0
pluggy==1.6.0
psycopg==3.2.13
psycopg-binary==3.2.13
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-discovery==1.2.0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.3
ruamel.yaml==0.19.1
s3transfer==0.16.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.48
tablib==3.9.0
tenacity==9.1.2
tomli==2.4.1
tomlkit==0.14.0
tqdm==4.67.3
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==21.2.0
Werkzeug==3.1.7
WTForms==3.2.1
zipp==3.23.0
